requirements.txt
opencv-python
numpy
opencolorio
requests
